Bug fixes in the Iele Compiler and update execution tests (#247)
* Bug fix: ecrecover returns 0 on error instead of revert.

* Updated failing execution tests: semanticTests/ecrecover.

* Updated failing execution tests: semanticTests/{array, calldata}.

* Updated failind execution tests: semanticTests/array.

* Updated failind execution tests: semanticTests/smoke.

* Updated failind execution tests: semanticTests/structs.

* Bug fix: computation of storage size of array element when the element is
recursive struct now returns 1 (i.e. size of pointer).

* Bug fix: nested structs now use the data location of enclosing structure,
in decode.

* Bug fix (cont'd): nested structs now use the data location of enclosing
structure, in encode.

* Update failing execution tests: semanticTests/structs/calldata.

* Updated failing compilation tests.

+++ test/libsolidity/semanticTests/array/calldata_array_of_struct.sol
@@ -10,7 +10,7 @@ contract C {
     function f(S[] calldata s)
         external
         pure
-        returns (uint256 l, uint256 a, uint256 b, uint256 c, uint256 d)
+        returns (uint l, uint256 a, uint256 b, uint256 c, uint256 d)
     {
         l = s.length;
         a = s[0].a;
@@ -24,4 +24,4 @@ contract C {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// f((uint256,uint256)[]): 0x20, 0x2, 0x1, 0x2, 0x3, 0x4 -> 2, 1, 2, 3, 4
+// f((uint256,uint256)[]): refargs { 0x01, 0x02, 0x0000000000000000000000000000000000000000000000000000000000000001, 0x0000000000000000000000000000000000000000000000000000000000000002, 0x0000000000000000000000000000000000000000000000000000000000000003, 0x0000000000000000000000000000000000000000000000000000000000000004 } -> 2, 1, 2, 3, 4
